Clean up facilityUser specific code into the facilityUserResource.
