+++ kolibri/core/assets/src/apiResources/facilityUser.js
@@ -1,9 +1,29 @@
 const Resource = require('../api_resource').Resource;
+const logging = require('logging').getLogger(__filename);
 
 class FacilityUserResource extends Resource {
   static resourceName() {
     return 'facilityuser';
   }
+
+  getCurrentFacility() {
+    const promise = new Promise((resolve, reject) => {
+      this.resource.client({ path: this.currentFacilityUrl() }).then((response) => {
+        resolve(response.entity);
+      }, (response) => {
+        logging.error('An error occurred', response);
+      });
+    },
+    (reject) => {
+      reject(reject);
+    });
+    return promise;
+  }
+
+  get currentFacilityUrl() {
+    return this.urls[`currentfacility_list`];
+  }
+
 }
 
 module.exports = FacilityUserResource;
